feat: refactor member domain structure and update imports for clarity
