+++ application/src/main/java/com/callv2/member/application/member/activation/DefaultTogleMemberActivationUseCase.java
@@ -3,9 +3,9 @@ package com.callv2.member.application.member.activation;
 import java.util.Objects;
 
 import com.callv2.member.domain.exception.NotFoundException;
-import com.callv2.member.domain.member.Member;
-import com.callv2.member.domain.member.MemberGateway;
-import com.callv2.member.domain.member.MemberID;
+import com.callv2.member.domain.member.entity.Member;
+import com.callv2.member.domain.member.entity.MemberID;
+import com.callv2.member.domain.member.gateway.MemberGateway;
 
 public class DefaultTogleMemberActivationUseCase extends TogleMemberActivationUseCase {
