+++ application/src/main/java/com/callv2/member/application/member/create/DefaultCreateMemberUseCase.java
@@ -4,13 +4,13 @@ import java.util.Objects;
 
 import com.callv2.member.domain.event.EventDispatcher;
 import com.callv2.member.domain.exception.DomainException;
-import com.callv2.member.domain.member.Email;
-import com.callv2.member.domain.member.Member;
-import com.callv2.member.domain.member.MemberGateway;
-import com.callv2.member.domain.member.Nickname;
-import com.callv2.member.domain.member.Password;
-import com.callv2.member.domain.member.PreMember;
-import com.callv2.member.domain.member.Username;
+import com.callv2.member.domain.member.entity.Member;
+import com.callv2.member.domain.member.gateway.MemberGateway;
+import com.callv2.member.domain.member.valueobject.Email;
+import com.callv2.member.domain.member.valueobject.Nickname;
+import com.callv2.member.domain.member.valueobject.Password;
+import com.callv2.member.domain.member.valueobject.PreMember;
+import com.callv2.member.domain.member.valueobject.Username;
 import com.callv2.member.domain.validation.handler.Notification;
 
 public class DefaultCreateMemberUseCase extends CreateMemberUseCase {
