+++ application/src/main/java/com/callv2/member/application/member/retrieve/get/DefaultGetMemberUseCase.java
@@ -3,9 +3,9 @@ package com.callv2.member.application.member.retrieve.get;
 import java.util.Objects;
 
 import com.callv2.member.domain.exception.NotFoundException;
-import com.callv2.member.domain.member.Member;
-import com.callv2.member.domain.member.MemberGateway;
-import com.callv2.member.domain.member.MemberID;
+import com.callv2.member.domain.member.entity.Member;
+import com.callv2.member.domain.member.entity.MemberID;
+import com.callv2.member.domain.member.gateway.MemberGateway;
 
 public class DefaultGetMemberUseCase extends GetMemberUseCase {
