+++ application/src/main/java/com/callv2/member/application/member/retrieve/get/GetMemberOutput.java
@@ -2,7 +2,7 @@ package com.callv2.member.application.member.retrieve.get;
 
 import java.time.Instant;
 
-import com.callv2.member.domain.member.Member;
+import com.callv2.member.domain.member.entity.Member;
 
 public record GetMemberOutput(
         String id,
