+++ domain/src/main/java/com/callv2/member/domain/member/entity/Member.java
@@ -1,4 +1,4 @@
-package com.callv2.member.domain.member;
+package com.callv2.member.domain.member.entity;
 
 import java.time.Instant;
 import java.util.Optional;
@@ -7,6 +7,11 @@ import java.util.Queue;
 import com.callv2.member.domain.AggregateRoot;
 import com.callv2.member.domain.event.Event;
 import com.callv2.member.domain.event.EventSource;
+import com.callv2.member.domain.member.event.MemberCreatedEvent;
+import com.callv2.member.domain.member.validation.MemberValidator;
+import com.callv2.member.domain.member.valueobject.Email;
+import com.callv2.member.domain.member.valueobject.Nickname;
+import com.callv2.member.domain.member.valueobject.Username;
 import com.callv2.member.domain.validation.ValidationHandler;
 
 public class Member extends AggregateRoot<MemberID> implements EventSource {
