+++ domain/src/main/java/com/callv2/member/domain/member/gateway/MemberGateway.java
@@ -1,7 +1,10 @@
-package com.callv2.member.domain.member;
+package com.callv2.member.domain.member.gateway;
 
 import java.util.Optional;
 
+import com.callv2.member.domain.member.entity.Member;
+import com.callv2.member.domain.member.entity.MemberID;
+import com.callv2.member.domain.member.valueobject.PreMember;
 import com.callv2.member.domain.pagination.Page;
 import com.callv2.member.domain.pagination.SearchQuery;
