+++ infrastructure/src/main/java/com/callv2/member/infrastructure/member/event/MemberCreatedEventMessagingHandler.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 
 import com.callv2.member.domain.event.Event;
 import com.callv2.member.domain.event.EventHandler;
-import com.callv2.member.domain.member.MemberCreatedEvent;
+import com.callv2.member.domain.member.event.MemberCreatedEvent;
 import com.callv2.member.infrastructure.messaging.producer.Producer;
 
 @Component
